+++ src/components/modals/ModalAlert.tsx
@@ -53,11 +53,12 @@ interface WithSecondaryButtonLabel extends BaseProps {
 
 export type ModalAlertProps = ModalCommonProps & (WithSecondaryButton | WithSecondaryButtonLabel);
 
-export const ModalAlert = forwardRef<HTMLButtonElement, ModalAlertProps>((props, ref) => {
+export const ModalAlert = forwardRef<HTMLDivElement, ModalAlertProps>((props, ref) => {
   const {
     children,
     description,
     content,
+    isDefaultOpen,
     isOpen,
     onOpenChange,
     title,
@@ -86,17 +87,15 @@ export const ModalAlert = forwardRef<HTMLButtonElement, ModalAlertProps>((props,
   }, []);
 
   return (
-    <RadixAlertDialog.Root onOpenChange={onOpenChange} open={isOpen}>
+    <RadixAlertDialog.Root defaultOpen={isDefaultOpen} onOpenChange={onOpenChange} open={isOpen}>
       <RadixAlertDialog.Trigger
         asChild
-        ref={ref}
-        {...(isDropdownMenuItemComponent(children) && { onSelect: handleSelectForDropdownMenu })}
-        {...radixProps}>
+        {...(isDropdownMenuItemComponent(children) && { onSelect: handleSelectForDropdownMenu })}>
         {children}
       </RadixAlertDialog.Trigger>
       <RadixAlertDialog.Portal>
         <ModalAlertOverlay />
-        <ModalAlertWrapper size={ModalSize.Default}>
+        <ModalAlertWrapper ref={ref} size={ModalSize.Default} {...radixProps}>
           <ModalAlertTitle>{title}</ModalAlertTitle>
 
           <ModalBody>
